Move DiscriminatedUnionDispatcher to shared location
