+++ experimental/dds/tree2/src/codec/index.ts
@@ -18,4 +18,5 @@ export {
 	withDefaultBinaryEncoding,
 	withSchemaValidation,
 } from "./codec";
+export { DiscriminatedUnionDispatcher, unionOptions } from "./discriminatedUnions";
 export { noopValidator } from "./noopValidator";
