+++ experimental/dds/tree2/src/feature-libraries/chunked-forest/codec/chunkDecodingGeneric.ts
@@ -5,15 +5,10 @@
 
 import { assert } from "@fluidframework/core-utils";
 import { BrandedType } from "../../../util";
+import { DiscriminatedUnionDispatcher } from "../../../codec";
 import { TreeChunk } from "../chunk";
 import { EncodedChunkGeneric, IdentifierOrIndex } from "./formatGeneric";
-import {
-	ChunkDecoder,
-	DiscriminatedUnionDispatcher,
-	StreamCursor,
-	getChecked,
-	readStream,
-} from "./chunkCodecUtilities";
+import { ChunkDecoder, StreamCursor, getChecked, readStream } from "./chunkCodecUtilities";
 
 /**
  * General purpose shape based tree decoder which gets its support for specific shapes from the caller.
