+++ experimental/dds/tree2/src/feature-libraries/chunked-forest/codec/formatGeneric.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ObjectOptions, Static, TSchema, Type } from "@sinclair/typebox";
+import { Static, TSchema, Type } from "@sinclair/typebox";
 
 /**
  * Identifier OR Index of an identifier in the identifier list.
@@ -23,17 +23,6 @@ export const ShapeIndex = Type.Number({ multipleOf: 1, minimum: 0 });
 export type ShapeIndex = Static<typeof ShapeIndex>;
 
 export const Count = Type.Number({ multipleOf: 1, minimum: 0 });
-
-/**
- * Options to configure a TypeBox schema as a discriminated union that is simple to validate data against.
- *
- * See DiscriminatedUnionDispatcher for more information on this pattern.
- */
-export const unionOptions: ObjectOptions = {
-	additionalProperties: false,
-	minProperties: 1,
-	maxProperties: 1,
-};
 
 const EncodedChunkBase = Type.Object(
 	{
